+++ src/services/users.services.ts
@@ -1,6 +1,26 @@
 import prisma from '../database/index.database';
 // import type { User } from '../../node_modules/.prisma/client';
 
+//creating users/signup
+export const userSignUp = async (
+  email1: string,
+  password1: string,
+  name1: string,
+) => {
+  try {
+    const newUser = await prisma.user.create({
+      data: {
+        email: email1,
+        password: password1,
+        name: name1,
+      },
+    });
+    return newUser;
+  } catch (error) {
+    console.log(error);
+    throw error;
+  }
+};
 //show all users
 export const showAllUsers = async () => {
   try {
